split the expressions by datasource
