+++ jx_python/containers/list_usingPythonList.py
@@ -14,15 +14,15 @@ from __future__ import unicode_literals
 import itertools
 from collections import Mapping
 
+from mo_collections import UniqueIndex
 from mo_dots import Data, wrap, listwrap, unwraplist, unwrap, Null
 from mo_logs import Log
 from mo_threads import Lock
 from pyLibrary import convert
 
-from mo_collections import UniqueIndex
+from jx_base.expressions import TRUE_FILTER, jx_expression, Expression, TrueOp, jx_expression_to_function, Variable
 from jx_python.containers import Container
 from jx_python.expression_compiler import compile_expression
-from jx_python.expressions import TRUE_FILTER, jx_expression, Expression, TrueOp, jx_expression_to_function, Variable
 from jx_python.lists.aggs import is_aggs, list_aggs
 from jx_python.meta import get_schema_from_list
