+++ jx_python/lists/aggs.py
@@ -21,7 +21,7 @@ from mo_collections.matrix import Matrix
 from jx_python import windows
 from jx_python.domains import SimpleSetDomain, DefaultDomain
 from jx_python.expression_compiler import compile_expression
-from jx_python.expressions import jx_expression_to_function
+from jx_base.expressions import jx_expression_to_function
 
 _ = Date
