+++ tests/test_jx/test_expressions.py
@@ -14,7 +14,7 @@ from __future__ import unicode_literals
 from jx_base.queries import is_variable_name
 from mo_testing.fuzzytestcase import FuzzyTestCase
 from mo_times.dates import Date
-from jx_python.expressions import simplify_esfilter, jx_expression, USE_BOOL_MUST
+from jx_base.expressions import simplify_esfilter, jx_expression, USE_BOOL_MUST
 
 
 class TestExpressions(FuzzyTestCase):
